fix: scan on demand cooldowns
