+++ src/hooks/useStore.js
@@ -83,6 +83,7 @@ export const useStore = create(
       setPopups: (popups) => set({ popups }),
       motdIndex: 0,
       setMotdIndex: (motdIndex) => set({ motdIndex }),
+      scannerCooldown: 0,
     }),
     {
       name: 'local-state',
@@ -156,5 +157,4 @@ export const useStatic = create((set) => ({
   setResetFilters: (resetFilters) => set({ resetFilters }),
   extraUserFields: [],
   setExtraUserFields: (extraUserFields) => set({ extraUserFields }),
-  scannerCooldown: 0,
 }))
